+++ app/src/main/java/ru/netology/diploma/entity/PostEntity.kt
@@ -21,7 +21,7 @@ data class PostEntity(
     val link: String?,
     val mentionIds: List<String>?,
     val mentionedMe: Boolean,
-    val published: Int,
+    val published: Long,
     @Embedded
     var attachment: AttachmentEmbeddable?,
 ) {
